# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AffiliateDeals API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Offer {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  discountPercent?: number | null;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  currentPrice?: number | null;
  currency?: string;
  /** @nullable */
  imageUrl?: string | null;
  affiliateUrl?: string;
  isActive: boolean;
  isFeatured: boolean;
  clickCount: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brandId?: number | null;
  /** @nullable */
  categorySlug?: string | null;
  /** @nullable */
  brandSlug?: string | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brandName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OfferDetail {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  shortDescription?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  discountPercent?: number | null;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  currentPrice?: number | null;
  currency?: string;
  /** @nullable */
  imageUrl?: string | null;
  affiliateUrl?: string;
  isActive: boolean;
  isFeatured: boolean;
  clickCount: number;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brandId?: number | null;
  /** @nullable */
  categorySlug?: string | null;
  /** @nullable */
  brandSlug?: string | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  brandName?: string | null;
  /** @nullable */
  brandLogoUrl?: string | null;
  relatedOffers?: Offer[];
  createdAt: string;
  updatedAt: string;
}

export interface OfferListResponse {
  items: Offer[];
  total: number;
  page: number;
  limit: number;
}

export interface OfferInput {
  /** @minLength 1 */
  title: string;
  slug?: string;
  shortDescription?: string;
  description?: string;
  discountPercent?: number;
  originalPrice?: number;
  currentPrice?: number;
  currency?: string;
  imageUrl?: string;
  /** @minLength 1 */
  affiliateUrl: string;
  isActive?: boolean;
  isFeatured?: boolean;
  categoryId?: number;
  brandId?: number;
}

export interface OfferUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  shortDescription?: string;
  description?: string;
  discountPercent?: number;
  originalPrice?: number;
  currentPrice?: number;
  currency?: string;
  imageUrl?: string;
  affiliateUrl?: string;
  isActive?: boolean;
  isFeatured?: boolean;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  brandId?: number | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  offerCount?: number;
  isActive: boolean;
}

export interface CategoryDetail {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isActive: boolean;
  offers: OfferListResponse;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export interface CategoryUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  description?: string;
  imageUrl?: string;
  isActive?: boolean;
}

export interface Brand {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  offerCount?: number;
  isActive: boolean;
}

export interface BrandDetail {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  isActive: boolean;
  offers: OfferListResponse;
}

export interface BrandInput {
  /** @minLength 1 */
  name: string;
  slug?: string;
  description?: string;
  logoUrl?: string;
  websiteUrl?: string;
  isActive?: boolean;
}

export interface BrandUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  description?: string;
  logoUrl?: string;
  websiteUrl?: string;
  isActive?: boolean;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  isPublished: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface BlogPostListResponse {
  items: BlogPost[];
  total: number;
  page: number;
  limit: number;
}

export interface BlogPostInput {
  /** @minLength 1 */
  title: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  imageUrl?: string;
  isPublished?: boolean;
}

export interface BlogPostUpdate {
  /** @minLength 1 */
  title?: string;
  slug?: string;
  excerpt?: string;
  content?: string;
  imageUrl?: string;
  isPublished?: boolean;
}

export type SearchResultItemType = typeof SearchResultItemType[keyof typeof SearchResultItemType];


export const SearchResultItemType = {
  offer: 'offer',
  category: 'category',
  brand: 'brand',
  blog: 'blog',
} as const;

export interface SearchResultItem {
  id: number;
  type: SearchResultItemType;
  title: string;
  slug: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface SearchResults {
  items: SearchResultItem[];
  total: number;
  page: number;
  query: string;
}

export interface AdminStats {
  totalOffers: number;
  activeOffers: number;
  featuredOffers: number;
  totalCategories: number;
  totalBrands: number;
  totalBlogPosts: number;
  publishedBlogPosts: number;
  totalClicks: number;
  pendingIndexing: number;
  failedIndexing: number;
  topOffers: Offer[];
  recentOffers: Offer[];
}

export type IndexingQueueItemStatus = typeof IndexingQueueItemStatus[keyof typeof IndexingQueueItemStatus];


export const IndexingQueueItemStatus = {
  pending: 'pending',
  sent: 'sent',
  failed: 'failed',
} as const;

export interface IndexingQueueItem {
  id: number;
  url: string;
  type: string;
  status: IndexingQueueItemStatus;
  provider: string;
  attempts: number;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  sentAt?: string | null;
}

export interface IndexingQueueResponse {
  items: IndexingQueueItem[];
  total: number;
  page: number;
}

export type ListOffersParams = {
page?: number;
limit?: number;
categorySlug?: string;
brandSlug?: string;
featured?: boolean;
q?: string;
};

export type TrackOfferClick200 = {
  affiliateUrl: string;
};

export type ListBlogPostsParams = {
page?: number;
limit?: number;
};

export type SearchParams = {
q: string;
page?: number;
};

export type ListIndexingQueueParams = {
status?: ListIndexingQueueStatus;
page?: number;
};

export type ListIndexingQueueStatus = typeof ListIndexingQueueStatus[keyof typeof ListIndexingQueueStatus];


export const ListIndexingQueueStatus = {
  pending: 'pending',
  sent: 'sent',
  failed: 'failed',
} as const;