# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AffiliateDeals API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List offers
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: categorySlug
          schema: { type: string }
        - in: query
          name: brandSlug
          schema: { type: string }
        - in: query
          name: featured
          schema: { type: boolean }
        - in: query
          name: q
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferListResponse"
  /offers/{slug}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get offer by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDetail"
        "404":
          description: Not found
  /offers/{slug}/track:
    post:
      operationId: trackOfferClick
      tags: [offers]
      summary: Track affiliate link click
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [affiliateUrl]
                properties:
                  affiliateUrl: { type: string }

  # ── Public: Categories ────────────────────────────────────────────────────
...
  /categories/{slug}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get category info
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found

  # ── Public: Brands ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Offer:
      type: object
      required: [id, title, slug, isActive, isFeatured, clickCount, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        shortDescription: { type: ["string", "null"] }
        discountPercent: { type: ["integer", "null"] }
        originalPrice: { type: ["number", "null"] }
        currentPrice: { type: ["number", "null"] }
        currency: { type: string, default: "USD" }
        imageUrl: { type: ["string", "null"] }
        affiliateUrl: { type: string }
        isActive: { type: boolean }
        isFeatured: { type: boolean }
        clickCount: { type: integer }
        categoryId: { type: ["integer", "null"] }
        brandId: { type: ["integer", "null"] }
        categorySlug: { type: ["string", "null"] }
        brandSlug: { type: ["string", "null"] }
        categoryName: { type: ["string", "null"] }
        brandName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OfferDetail:
      type: object
      required: [id, title, slug, isActive, isFeatured, clickCount, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        shortDescription: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        discountPercent: { type: ["integer", "null"] }
        originalPrice: { type: ["number", "null"] }
        currentPrice: { type: ["number", "null"] }
        currency: { type: string }
        imageUrl: { type: ["string", "null"] }
        affiliateUrl: { type: string }
        isActive: { type: boolean }
        isFeatured: { type: boolean }
        clickCount: { type: integer }
        categoryId: { type: ["integer", "null"] }
        brandId: { type: ["integer", "null"] }
        categorySlug: { type: ["string", "null"] }
        brandSlug: { type: ["string", "null"] }
        categoryName: { type: ["string", "null"] }
        brandName: { type: ["string", "null"] }
        brandLogoUrl: { type: ["string", "null"] }
        relatedOffers:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    OfferListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Offer"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    Category:
      type: object
      required: [id, name, slug, isActive]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        offerCount: { type: integer, default: 0 }
        isActive: { type: boolean }
...